Legge inn fho i print
