+++ src/moduler/utskrift/print/aktiviteter/ForhaandsorienteringPrint.module.less
@@ -1,0 +1,7 @@
+.forhaandsorientering {
+    margin: 0.3rem 0;
+}
+
+.lestTekst {
+    font-style: italic;
+}
